refactor: extract the Cognito trigger invocation
Reaching the function a trigger names moves into
SimCognitoTriggerInvocation, leaving SimCognitoUserPoolTriggers as the
method per trigger that says when it fires and what the pool does with
the answer. The fifth trigger had taken that class past the complexity
threshold, and the two halves read on their own.

Part of #427.
